Changing deploy flow
