+++ docker-compose.prod.yml
@@ -1,30 +1,53 @@
-version: '3.8'
+# 1) build with tag commit
+#TAG=$(git rev-parse --short HEAD) docker compose -f docker-compose.yml -f docker-compose.prod.yml build
+
+# 2) release tasks
+#AG=$(git rev-parse --short HEAD) docker compose -f docker-compose.yml -f docker-compose.prod.yml run --rm release
+
+# 3) switch images
+#TAG=$(git rev-parse --short HEAD) docker compose -f docker-compose.yml -f docker-compose.prod.yml up -d
+
 
 services:
+  # ---------- Frontend ----------
   portfolio-fe:
     build:
       context: ./frontend
       target: prod
+    image: portfolio-frontend:${TAG:-latest}
     restart: always
-    # Remove development volumes to use the optimized built artifacts
+    # No dev volumes in prod
     volumes:
       - ./nginx/ssl/certs:/etc/ssl/certs:ro
 
+  # ---------- Backend ----------
   portfolio-be:
     build:
+      context: ./backend
       target: production
+    image: portfolio-backend:${TAG:-latest}
     restart: always
+    # No code mounts in prod
+    volumes:
+      # persist user uploads + static collected assets
+      - backend_media:/app/media
+      - backend_static:/app/staticfiles
     environment:
       DEBUG: "False"
-      # Core settings injected via Doppler/Shell
+      DOCKER_ENV: "true"
+      PYTHONPATH: "."
+
+      # Core
       SECRET_KEY: ${SECRET_KEY}
+
       # Database
       DB_ENGINE: ${DB_ENGINE}
       DB_NAME: ${DB_NAME}
       DB_USER: ${DB_USER}
       DB_PASSWORD: ${DB_PASSWORD}
-      DB_HOST: ${DB_HOST}
-      DB_PORT: ${DB_PORT}
+      DB_HOST: ${DB_HOST:-db}
+      DB_PORT: ${DB_PORT:-5432}
+
       # Domain settings
       ADMIN_DOMAIN: ${ADMIN_DOMAIN}
       API_DOMAIN: ${API_DOMAIN}
@@ -32,6 +55,7 @@ services:
       ALLOWED_HOSTS: ${ALLOWED_HOSTS}
       CORS_ALLOWED_ORIGINS: ${CORS_ALLOWED_ORIGINS}
       CSRF_TRUSTED_ORIGINS: ${CSRF_TRUSTED_ORIGINS}
+
       # Email
       EMAIL_BACKEND: ${EMAIL_BACKEND}
       EMAIL_HOST: ${EMAIL_HOST}
@@ -42,8 +66,66 @@ services:
       DEFAULT_FROM_EMAIL: ${DEFAULT_FROM_EMAIL}
       CONTACT_EMAIL: ${CONTACT_EMAIL}
 
+  # ✅ Release job (runs once per deploy)
+  release:
+    image: portfolio-backend:${TAG:-latest}
+    restart: "no"
+    depends_on:
+      - db
+    volumes:
+      - backend_media:/app/media
+      - backend_static:/app/staticfiles
+    environment:
+      DEBUG: "False"
+      DOCKER_ENV: "true"
+      PYTHONPATH: "."
+
+      SECRET_KEY: ${SECRET_KEY}
+
+      DB_ENGINE: ${DB_ENGINE}
+      DB_NAME: ${DB_NAME}
+      DB_USER: ${DB_USER}
+      DB_PASSWORD: ${DB_PASSWORD}
+      DB_HOST: ${DB_HOST:-db}
+      DB_PORT: ${DB_PORT:-5432}
+
+      ADMIN_DOMAIN: ${ADMIN_DOMAIN}
+      API_DOMAIN: ${API_DOMAIN}
+      CSRF_COOKIE_DOMAIN: ${CSRF_COOKIE_DOMAIN}
+      ALLOWED_HOSTS: ${ALLOWED_HOSTS}
+      CORS_ALLOWED_ORIGINS: ${CORS_ALLOWED_ORIGINS}
+      CSRF_TRUSTED_ORIGINS: ${CSRF_TRUSTED_ORIGINS}
+
+      EMAIL_BACKEND: ${EMAIL_BACKEND}
+      EMAIL_HOST: ${EMAIL_HOST}
+      EMAIL_PORT: ${EMAIL_PORT}
+      EMAIL_USE_TLS: ${EMAIL_USE_TLS}
+      EMAIL_HOST_USER: ${EMAIL_HOST_USER}
+      EMAIL_HOST_PASSWORD: ${EMAIL_HOST_PASSWORD}
+      DEFAULT_FROM_EMAIL: ${DEFAULT_FROM_EMAIL}
+      CONTACT_EMAIL: ${CONTACT_EMAIL}
+
+    command: >
+      sh -c "
+        python manage.py migrate --noinput &&
+        python manage.py compilemessages &&
+        python manage.py collectstatic --noinput
+      "
+
+  # ---------- Nginx ----------
   portfolio-nginx:
     restart: always
+    # nginx container needs access to media/static produced by backend/release
+    volumes:
+      - ./nginx/nginx.conf:/etc/nginx/nginx.conf:ro
+      - ./nginx/ssl/certs:/etc/nginx/ssl:ro
+      - ./nginx/logs:/var/log/nginx
+      - backend_media:/app/media:ro
+      - backend_static:/app/staticfiles:ro
     ports:
       - "80:80"
       - "443:443"
+
+volumes:
+  backend_media:
+  backend_static:
