+++ docker-compose.yml
@@ -1,9 +1,5 @@
 networks:
   portfolio_network:
-    driver: bridge
-    ipam:
-      config:
-        - subnet: 172.10.0.0/16
 
 x-backend-common: &backend-common
   build:
@@ -17,7 +13,7 @@ x-backend-common: &backend-common
     - path: ./backend/.env
       required: false
   environment:
-    DOCKER_ENV: true
+    DOCKER_ENV: "true"
     PYTHONPATH: "."
   depends_on:
     - db
@@ -29,40 +25,31 @@ services:
       context: ./frontend
       target: dev
     networks:
-      portfolio_network:
-        ipv4_address: 172.10.0.2
+      - portfolio_network
     ports:
       - "3000:3000"
     restart: unless-stopped
     volumes:
-      - ./frontend/src:/app/src
-      - ./frontend/public:/app/public
-      - ./frontend/package.json:/app/package.json
-      - ./frontend/webpack.config.js:/app/webpack.config.js
-      - ./frontend/.babelrc:/app/.babelrc
-      - ./frontend/node_modules/@types:/app/node_modules/@types
-      - ./nginx/ssl/certs:/etc/ssl/certs:ro
+      - ./frontend:/app
+      - fe_node_modules:/app/node_modules
+    # jeśli potrzebujesz certów lokalnie:
+    # - ./nginx/ssl/certs:/etc/ssl/certs:ro
 
   portfolio-be:
     <<: *backend-common
-    networks:
-      portfolio_network:
-        ipv4_address: 172.10.0.3
     ports:
       - "8000:8000"
 
   portfolio-test:
     <<: *backend-common
-    profiles:
-      - test
+    profiles: ["test"]
     command: pytest
 
   db:
     image: postgres:15
     restart: unless-stopped
     networks:
-      portfolio_network:
-        ipv4_address: 172.10.0.4
+      - portfolio_network
     env_file:
       - path: ./backend/.env
         required: false
@@ -79,16 +66,14 @@ services:
     image: redis:7-alpine
     restart: unless-stopped
     networks:
-      portfolio_network:
-        ipv4_address: 172.10.0.6
+      - portfolio_network
     ports:
       - "6379:6379"
 
   portfolio-nginx:
     image: nginx:1.25
     networks:
-      portfolio_network:
-        ipv4_address: 172.10.0.5
+      - portfolio_network
     depends_on:
       - portfolio-fe
       - portfolio-be
@@ -103,3 +88,4 @@ services:
 
 volumes:
   postgres_data:
+  fe_node_modules:
