+++ frontend/Dockerfile
@@ -1,26 +1,21 @@
-# --- Base image ---
+# --- Base deps ---
 FROM node:20-alpine AS base
 WORKDIR /app
 COPY package.json package-lock.json ./
 RUN npm ci
 
-# --- Development image ---
+# --- Development ---
 FROM base AS dev
 COPY . .
-EXPOSE 3000
 RUN chown -R node:node /app
 USER node
+EXPOSE 3000
 CMD ["npm", "start"]
 
 # --- Production build ---
 FROM base AS build
 COPY . .
 RUN npm run build
-
-# --- Test image ---
-FROM base AS test
-COPY . .
-CMD ["npm", "test", "--", "--passWithNoTests"]
 
 # --- Production runtime ---
 FROM node:20-alpine AS prod
@@ -30,11 +25,3 @@ USER node
 COPY --from=build /app/dist ./dist
 EXPOSE 3000
 CMD ["serve", "-s", "dist", "-l", "3000"]
-
-# --- Final stage switcher ---
-ARG NODE_ENV=development
-FROM dev AS final-dev
-FROM prod AS final-prod
-
-# Use the right final stage
-# docker build --target final-prod --build-arg NODE_ENV=production .
