+++ frontend/e2e/travel-highlights.spec.ts
@@ -1,80 +1,34 @@
-import { test, expect } from '@playwright/test';
+import { test, expect } from './fixtures';
 
 test.describe('Travel Highlights Page', () => {
   test.beforeEach(async ({ page }) => {
-    // Catch-all mock for API v1
-    await page.route('**/api/v1/**', async route => {
-      const url = route.request().url();
-
-      if (url.includes('/profile/')) {
-        return route.fulfill({
-          status: 200,
-          contentType: 'application/json',
-          body: JSON.stringify({
-            first_name: 'Test',
-            last_name: 'User',
-            short_description: 'Sky Hunter',
-          }),
-        });
-      }
-
-      if (url.includes('/background/')) {
-        return route.fulfill({
-          status: 200,
-          contentType: 'application/json',
-          body: JSON.stringify({
-            url: 'https://via.placeholder.com/1920x1080',
-          }),
-        });
-      }
-
-      if (url.includes('/whats-enabled/')) {
-        return route.fulfill({
-          status: 200,
-          contentType: 'application/json',
-          body: JSON.stringify({
-            travelHighlights: true,
-            gallery: true,
-            contactForm: true,
-            lastimages: true,
-          }),
-        });
-      }
-
-      if (url.includes('/travel/iceland/')) {
-        return route.fulfill({
-          status: 200,
-          contentType: 'application/json',
-          body: JSON.stringify({
-            country: 'Iceland',
-            place: null,
-            story: '<p>A land of fire and ice...</p>',
-            adventure_date: 'Jan 2026',
-            created_at: '2026-01-01T00:00:00Z',
-            highlight_name: 'Iceland Expedition',
-            background_image: 'https://via.placeholder.com/1920x1080',
-            images: [
-              {
-                pk: 101,
-                name: 'Northern Lights',
-                url: 'https://via.placeholder.com/800x600',
-                thumbnail_url: 'https://via.placeholder.com/200x150',
-                description: 'Dancing lights.',
-                location: 'Reykjavik',
-                camera: [{ model: 'Canon R6' }],
-                lens: [{ model: '14mm F1.8' }],
-                telescope: [],
-              },
-            ],
-          }),
-        });
-      }
-
-      // Default fallback
+    // Override specific endpoint for travel detail
+    await page.route('**/api/v1/travel/iceland/', async route => {
       return route.fulfill({
         status: 200,
         contentType: 'application/json',
-        body: JSON.stringify([]),
+        body: JSON.stringify({
+          country: 'Iceland',
+          place: null,
+          story: '<p>A land of fire and ice...</p>',
+          adventure_date: 'Jan 2026',
+          created_at: '2026-01-01T00:00:00Z',
+          highlight_name: 'Iceland Expedition',
+          background_image: 'https://via.placeholder.com/1920x1080',
+          images: [
+            {
+              pk: 101,
+              name: 'Northern Lights',
+              url: 'https://via.placeholder.com/800x600',
+              thumbnail_url: 'https://via.placeholder.com/200x150',
+              description: 'Dancing lights.',
+              location: 'Reykjavik',
+              camera: [{ model: 'Canon R6' }],
+              lens: [{ model: '14mm F1.8' }],
+              telescope: [],
+            },
+          ],
+        }),
       });
     });
