+++ nginx/nginx.conf
@@ -58,8 +58,7 @@ http {
         error_log /var/log/nginx/portfolio-ssl.error.log notice;
 
         location / {
-            proxy_pass https://frontend;
-            proxy_ssl_verify off;
+            proxy_pass http://frontend;
             proxy_http_version 1.1;
             proxy_set_header Upgrade $http_upgrade;
             proxy_set_header Connection 'upgrade';
