+++ nginx/prod/nginx.conf
@@ -1,0 +1,160 @@
+# ============================================================
+# PROD CONFIG VARIABLES (CHANGE THESE ON PRODUCTION)
+# ============================================================
+
+# Domains
+set $FRONTEND_DOMAIN example.com;
+set $FRONTEND_DOMAIN_WWW www.example.com;
+set $API_DOMAIN api.example.com;
+set $ADMIN_DOMAIN admin.example.com;
+
+# SSL certificate paths
+set $SSL_CERT     /etc/nginx/ssl/fullchain.pem;
+set $SSL_CERT_KEY /etc/nginx/ssl/privkey.pem;
+
+# Upstream services (Docker service names)
+set $FRONTEND_UPSTREAM http://portfolio-fe:3000;
+set $BACKEND_UPSTREAM  http://portfolio-be:8000;
+
+# Max upload size
+client_max_body_size 50M;
+
+# ============================================================
+
+worker_processes auto;
+
+events {
+    worker_connections 1024;
+}
+
+http {
+    include       /etc/nginx/mime.types;
+    default_type  application/octet-stream;
+
+    sendfile on;
+    keepalive_timeout 65;
+
+    resolver 127.0.0.11 valid=30s;
+
+    # -------------------------
+    # FRONTEND DOMAIN
+    # -------------------------
+
+    server {
+        listen 80;
+        server_name $FRONTEND_DOMAIN $FRONTEND_DOMAIN_WWW;
+        return 301 https://$host$request_uri;
+    }
+
+    server {
+        listen 443 ssl;
+        server_name $FRONTEND_DOMAIN $FRONTEND_DOMAIN_WWW;
+
+        ssl_certificate     $SSL_CERT;
+        ssl_certificate_key $SSL_CERT_KEY;
+
+        ssl_protocols TLSv1.2 TLSv1.3;
+        ssl_prefer_server_ciphers on;
+
+        access_log /var/log/nginx/frontend.access.log;
+        error_log  /var/log/nginx/frontend.error.log notice;
+
+        location / {
+            proxy_pass $FRONTEND_UPSTREAM;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+        }
+
+        # OPTIONAL: API under frontend domain
+        location /api/ {
+            proxy_pass $BACKEND_UPSTREAM;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+            proxy_read_timeout 120s;
+        }
+    }
+
+    # -------------------------
+    # API DOMAIN
+    # -------------------------
+
+    server {
+        listen 80;
+        server_name $API_DOMAIN;
+        return 301 https://$host$request_uri;
+    }
+
+    server {
+        listen 443 ssl;
+        server_name $API_DOMAIN;
+
+        ssl_certificate     $SSL_CERT;
+        ssl_certificate_key $SSL_CERT_KEY;
+
+        ssl_protocols TLSv1.2 TLSv1.3;
+        ssl_prefer_server_ciphers on;
+
+        access_log /var/log/nginx/api.access.log;
+        error_log  /var/log/nginx/api.error.log notice;
+
+        location / {
+            proxy_pass $BACKEND_UPSTREAM;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+            proxy_read_timeout 120s;
+        }
+    }
+
+    # -------------------------
+    # ADMIN DOMAIN
+    # -------------------------
+
+    server {
+        listen 80;
+        server_name $ADMIN_DOMAIN;
+        return 301 https://$host$request_uri;
+    }
+
+    server {
+        listen 443 ssl;
+        server_name $ADMIN_DOMAIN;
+
+        ssl_certificate     $SSL_CERT;
+        ssl_certificate_key $SSL_CERT_KEY;
+
+        ssl_protocols TLSv1.2 TLSv1.3;
+        ssl_prefer_server_ciphers on;
+
+        access_log /var/log/nginx/admin.access.log;
+        error_log  /var/log/nginx/admin.error.log notice;
+
+        location / {
+            proxy_pass $BACKEND_UPSTREAM;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+            proxy_read_timeout 120s;
+        }
+
+        location /static/ {
+            alias /app/staticfiles/;
+            autoindex off;
+            expires 30d;
+            add_header Cache-Control "public";
+        }
+
+        location /media/ {
+            alias /app/media/;
+            autoindex off;
+            expires 30d;
+            add_header Cache-Control "public";
+        }
+    }
+}
